API functions updates and new example
API functions added which control the LMS8FE board and
specifically SC1905 chip on it. Minor bugs fixed in GUI.

An example has been created that makes linearization process up and running. The example relies on previously created .ini files for the LMS8001 ICs and the board itself.
These .ini files are loaded.